Move base Adyen module into an independent file
